requirements.txt
pdfplumber
pandas
pydantic
transformers
torch
sentence-transformers
chromadb
langchain
langchain-openai
tiktoken
fastapi
uvicorn
python-dotenv
